feat: add notification navigation
