+++ app/src/androidTest/java/com/android/wildex/ui/home/HomeScreenTest.kt
@@ -573,7 +573,7 @@ class HomeScreenTest {
   @Test
   fun postImageAndCommentCountClick_invokesCallback() {
     fakeObserver.setOnline(true)
-    var postClicked = false
+    var postClicked: Boolean
     runBlocking {
       postRepository.addPost(fullPost)
       homeScreenVM.refreshUIState()
