+++ app/src/main/java/com/android/wildex/ui/profile/EditProfileTopBar.kt
@@ -1,8 +1,5 @@
 package com.android.wildex.ui.profile
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ChevronLeft
 import androidx.compose.material3.CenterAlignedTopAppBar
@@ -13,11 +10,9 @@ import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.MaterialTheme.typography
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.unit.dp
+import androidx.compose.ui.res.stringResource
 import com.android.wildex.R
 
 /**
@@ -30,35 +25,26 @@ import com.android.wildex.R
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun EditProfileTopBar(isNewUser: Boolean, onGoBack: () -> Unit) {
-  val cs = colorScheme
-  val context = LocalContext.current
+fun EditProfileTopBar(onGoBack: () -> Unit) {
   CenterAlignedTopAppBar(
       title = {
-        Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
-          Text(
-              text =
-                  if (isNewUser) LocalContext.current.getString(R.string.create_profile_title)
-                  else LocalContext.current.getString(R.string.edit_profile_title),
-              style = typography.titleLarge,
-              color = cs.onBackground,
-          )
-        }
+        Text(
+            text = stringResource(R.string.edit_profile_title),
+            style = typography.titleLarge,
+            color = colorScheme.onBackground,
+        )
       },
       navigationIcon = {
         IconButton(
             modifier = Modifier.testTag(EditProfileScreenTestTags.GO_BACK),
-            onClick = { onGoBack() },
+            onClick = onGoBack,
         ) {
           Icon(
               imageVector = Icons.Default.ChevronLeft,
-              contentDescription = context.getString(R.string.back),
-              tint = cs.onBackground,
+              contentDescription = stringResource(R.string.back),
+              tint = colorScheme.onBackground,
           )
         }
       },
-      actions = {
-        // Empty box to center the title
-        Box(modifier = Modifier.size(48.dp))
-      })
+  )
 }
