+++ app/src/main/java/com/android/wildex/ui/utils/images/ImageWithDoubleTap.kt
@@ -20,6 +20,7 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.graphics.painter.ColorPainter
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
@@ -58,7 +59,11 @@ fun ImageWithDoubleTapLike(
             model = pictureURL,
             contentDescription = "Post picture",
             contentScale = ContentScale.Crop,
-            modifier = Modifier.fillMaxSize())
+            modifier = Modifier.fillMaxSize(),
+            error = ColorPainter(colorScheme.surface.copy(.7f)),
+            placeholder = ColorPainter(colorScheme.surface.copy(.7f)),
+            fallback = ColorPainter(colorScheme.surface.copy(.7f)),
+        )
 
         DoubleTapHeartOverlay(
             isVisible = isHeartOverlayVisible,
